fix test

+++ e2e-test/application_workflow_test.go
@@ -155,7 +155,7 @@ var _ = Describe("Test application workflow rest api", func() {
 		var lrr apisv1.ListWorkflowRecordsResponse
 		Expect(decodeResponseBody(res, &lrr)).Should(Succeed())
 		Expect(lrr.Total).Should(Equal(int64(1)))
-		Expect(len(lrr.Records)).Should(Equal(int64(1)))
+		Expect(len(lrr.Records)).Should(Equal(1))
 	})
 
 	It("Detail the record", func() {
